toyrobot.py: Respond appropriately to premature requests
It doesnt make sense to ask the robot to MOVE or TURN before it
has even been placed on the table. Similarly it should not
report any coordinates or direction before it has been placed
on the table. This change updates the logic to make use of the
new placed attribute to determine whether to refuse the command
(MOVE and TURN) or to reply with the status that it is ready
to be placed (REPORT).